Consolidating post-related formatting functions into /utils :art:
-Cleaning up code in blog components by factoring out common functions into /utils/postFormatter.js
